Bug 1499026 - Part 2: Update in-tree ICU to release 63.1. rs=Waldo
UltraBlame original commit: 9deb23f8b29b29a3606f9c9536b5c4da6e27f0e3

+++ intl/icu/source/common/bmpset.cpp
@@ -241,13 +241,13 @@ void BMPSet::overrideIllegal() {
             bmpBlockBits[i]|=bits;
         }
 
-        mask=~(0x10001<<0xd);   
+        mask= static_cast<uint32_t>(~(0x10001<<0xd));   
         bits=1<<0xd;
         for(i=32; i<64; ++i) {  
             bmpBlockBits[i]=(bmpBlockBits[i]&mask)|bits;
         }
     } else {
-        mask=~(0x10001<<0xd);   
+        mask= static_cast<uint32_t>(~(0x10001<<0xd));   
         for(i=32; i<64; ++i) {  
             bmpBlockBits[i]&=mask;
         }
