+++ intl/icu/source/tools/toolutil/unewdata.cpp
@@ -61,17 +61,17 @@ udata_create(const char *dir, const char *type, const char *name,
     length = 0;					
     if(dir != NULL  && *dir !=0)	
     {
-    	length += strlen(dir);
+    	length += static_cast<int32_t>(strlen(dir));
 	
     	
         if (dir[strlen(dir) - 1]!= U_FILE_SEP_CHAR) {
             length++;
         }
 	}
-    length += strlen(name);		
+    length += static_cast<int32_t>(strlen(name));		
 
     if(type != NULL  && *type !=0) { 
-        length += strlen(type);
+        length += static_cast<int32_t>(strlen(type));
     }
